// frontend/src/api-generated/models/loaiTaiSanResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface LoaiTaiSanResponse {
  id?: number;
  maLoai?: string;
  tenLoai?: string;
  tienToMaThe?: string;
  thoiGianKhauHao?: number;
  ghiChu?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}