// frontend/src/api-generated/models/loaiTaiSanRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface LoaiTaiSanRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  maLoai?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  tenLoai: string;
  /**
     * @minLength 0
     * @maxLength 10
     */
  tienToMaThe?: string;
  /** @minimum 0 */
  thoiGianKhauHao?: number;
  ghiChu?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}